// auth-service/src/modules/admin/admin-employee/dto/employee.dto.ts
export class EmployeeDto {
  employeeCode: string;
  employeeName: string;
  position: string;
  gender: string;
  address: string;
  email: string;
  phoneNumber: string;
  dateOfBirth: Date;
  startDate: Date;
  status: string;
  departmentId: number;
  username: string;
  password: string;
  role: string;
}